Add CLI struct
