+++ src/lib.rs
@@ -1,4 +1,5 @@
+pub mod cli;
 mod direction;
 mod merkle_hash;
-pub mod merkle_tree;
+mod merkle_tree;
 mod proof_of_inclusion;
